+++ cypress/integration/github.spec.ts
@@ -1,3 +1,5 @@
+import { cssPodUrl } from '@cy/support/commands/auth';
+
 describe('Github issues', () => {
 
     beforeEach(() => {
@@ -20,6 +22,31 @@ describe('Github issues', () => {
 
         // Assert
         cy.see('Change recipe image');
+    });
+
+    it('#14 "Back to cookbook" navigates to authorization page', () => {
+        // Arrange
+        cy.visit('/?authenticator=inrupt');
+        cy.startApp();
+        cy.press('Connect your Solid POD');
+        cy.ariaInput('Login url').clear().type(cssPodUrl('/alice/{enter}'));
+        cy.cssAuthorize({
+            reset: {
+                typeIndex: true,
+                cookbook: true,
+                recipe: 'ramen',
+            },
+        });
+        cy.waitForReload({ resetProfiles: true });
+        cy.press('Ramen');
+        cy.see('Broth');
+        cy.reload();
+
+        // Act
+        cy.press('back to cookbook');
+
+        // Assert
+        cy.url().should('equal', `${Cypress.config('baseUrl')}/`);
     });
 
 });